# internal/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /v2/server/{server_id}/links:
    get:
      description: Returns all player-Discord account links for server members.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      - description: Max results per page (default 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Search by player tag
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get server links
      tags:
      - Server Links
  /v2/server/{server_id}/links/{user_discord_id}/{player_tag}:
    delete:
      description: Removes the link between a Discord user and a player account.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      - description: Discord User ID
        in: path
        name: user_discord_id
        required: true
        type: string
      - description: Player Tag
        in: path
        name: player_tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a link
      tags:
      - Server Links
  /v2/server/{server_id}/links/bulk-unlink:
    post:
      consumes:
      - application/json
      description: Removes multiple player-Discord links for a user in bulk.
      parameters:
      - description: Server ID
        in: path
        name: server_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Bulk unlink accounts
      tags:
      - Server Links
...
swagger: "2.0"